requirements.txt
streamlit
pandas
numpy
pyodbc
sqlalchemy
watchdog
requests
openpyxl
plotly
bcrypt
python-dotenv
